handle deleted protocols, move helper to own file
